nginx up/down only on docker-host-01

+++ star.mfg.otaya.letsencrypt.update.sh
@@ -2,11 +2,18 @@
 
 set -e
 
-systemctl stop docker-compose-ephemeral@nginx-reverse-proxy.service
+# stop the reverse proxy nginx docker container
+if [[ "xdocker-host-01" -eq "x$HOSTNAME" ]]
+then
+  systemctl stop docker-compose-ephemeral@nginx-reverse-proxy.service
+fi
 
-. /home/andi/docker-server-env/nginx-reverse-proxy-certbot/aws-env
-. /home/andi/docker-server-env/nginx-reverse-proxy-certbot/domains-env
+# load the environment
+DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" >/dev/null 2>&1 && pwd )"
+. ${DIR}/aws-env
+. ${DIR}/domains-env
 
+# update the certificates
 docker run \
   -v nginx-certs:/etc/letsencrypt \
   -e http_proxy=$http_proxy \
@@ -14,8 +21,10 @@ docker run \
   -e email="${EMAIL}" \
   -e AWS_ACCESS_KEY_ID="$AWS_ACCESS_KEY_ID" \
   -e AWS_SECRET_ACCESS_KEY="$AWS_SECRET_ACCESS_KEY" \
-  -p 10.0.0.27:80:80 \
-  -p 10.0.0.27:443:443 \
   --rm ascheucher/dockerized-certbot:latest
 
-systemctl start docker-compose-ephemeral@nginx-reverse-proxy.service
+# start the reverse proxy nginx docker container again
+if [[ "xdocker-host-01" -eq "x$HOSTNAME" ]]
+then
+  systemctl start docker-compose-ephemeral@nginx-reverse-proxy.service
+fi
